+++ src/main/java/com/alexberemart/finances/domain/ports/dtos/ImportMovementDto.java
@@ -1,18 +1,19 @@
 package com.alexberemart.finances.domain.ports.dtos;
 
 import java.util.Date;
+import java.math.BigDecimal;
 
 public class ImportMovementDto {
 
     private Date date;
     private String description;
-    private Double amount;
+    private BigDecimal amount; // Changed from Double to BigDecimal
 
-    public Double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
